Break apart MyCollectionArtowrkDetails
