+++ src/lib/Scenes/MyCollection/Screens/AddArtwork/AddEditArtwork.tsx
@@ -13,18 +13,11 @@ import { MediumPicker } from "./Components/MediumPicker"
 export const AddEditArtwork: React.FC = () => {
   const artworkActions = AppStore.actions.myCollection.artwork
   const navActions = AppStore.actions.myCollection.navigation
-  const artworkState = AppStore.useAppState(state => state.myCollection.artwork)
-  const navState = AppStore.useAppState(state => state.myCollection.navigation)
+  const artworkState = AppStore.useAppState((state) => state.myCollection.artwork)
+  const navState = AppStore.useAppState((state) => state.myCollection.navigation)
   const { formik } = useArtworkForm()
   const photos = artworkState.sessionState.formValues.photos
   const modalType = navState?.sessionState?.modalType
-
-  // TODO: Do we still need this now that stuff has been moved to its own screen?
-  // const formattedTitleAndYear = [formik.values.title, formik.values.date].filter(Boolean).join(", ")
-  // const formattedDimensions = [formik.values.height, formik.values.width, formik.values.depth]
-  // .filter(Boolean)
-  // .join(" × ")
-
   const addOrEditLabel = modalType ? "Edit" : "Add"
 
   return (
