+++ src/lib/Scenes/MyCollection/Screens/ArtworkDetail/Components/MyCollectionArtworkMeta.tsx
@@ -1,6 +1,7 @@
 import { MyCollectionArtworkMeta_artwork } from "__generated__/MyCollectionArtworkMeta_artwork.graphql"
 import { CaretButton } from "lib/Components/Buttons/CaretButton"
 import { ScreenMargin } from "lib/Scenes/MyCollection/Components/ScreenMargin"
+import { formatArtworkDimensions } from "lib/Scenes/MyCollection/utils/formatArtworkDimensions"
 import { AppStore } from "lib/store/AppStore"
 import { capitalize } from "lodash"
 import { Spacer } from "palette"
@@ -15,43 +16,57 @@ interface MyCollectionArtworkMetaProps {
 
 const MyCollectionArtworkMeta: React.FC<MyCollectionArtworkMetaProps> = ({ artwork, viewAll = false }) => {
   const navActions = AppStore.actions.myCollection.navigation
+  const { artistNames, category, date, depth, height, medium, metric, title, width } = artwork
+  const dimensions = formatArtworkDimensions({ height, width, depth, metric })
 
-  return (
-    <ScreenMargin>
-      {/*
-        NOTE: `medium` prop is correct; catergory is the label. `materials` field
-        corresponds to `category` prop. It's a mess :/
-      */}
-      <Field label="Category" value={capitalize(artwork?.medium as string)} />
-      <Field label="Dimensions" value={capitalize(artwork?.dimensions?.in as string)} />
-      <Field label="Edition" value="TODO" />
+  if (viewAll) {
+    return (
+      <ScreenMargin>
+        <Field label="Artist" value={artistNames} />
+        <Field label="Title" value={title} />
+        <Field label="Year created" value={date} />
+        {/*
+          NOTE: `medium` prop is correct; catergory is the label. `materials` field
+          corresponds to `category` prop. It's a mess :/
+        */}
+        <Field label="Category" value={capitalize(medium as string)} />
+        <Field label="Materials" value={capitalize(category as string)} />
+        <Field label="Dimensions" value={dimensions} />
+        <Field label="Edition" value="TODO" />
+        <Field label="Price paid" value="TODO" />
+      </ScreenMargin>
+    )
+  } else {
+    return (
+      <ScreenMargin>
+        <Field label="Category" value={capitalize(medium as string)} />
+        <Field label="Dimensions" value={dimensions} />
+        <Field label="Edition" value="TODO" />
+        <Field label="Price paid" value="TODO" />
 
-      {!!viewAll && (
-        <>
-          <Field label="Title" value={artwork.title} />
-          <Field label="Year created" value={artwork.date} />
-        </>
-      )}
+        <Spacer my={0.5} />
 
-      <Spacer my={0.5} />
-
-      <CaretButton
-        onPress={() => navActions.navigateToViewAllArtworkDetails({ passProps: artwork })}
-        text="View more"
-      />
-    </ScreenMargin>
-  )
+        <CaretButton
+          onPress={() => navActions.navigateToViewAllArtworkDetails({ passProps: artwork })} // FIXME: need to fix NavigatorIOS
+          text="View more"
+        />
+      </ScreenMargin>
+    )
+  }
 }
 
 export const MyCollectionArtworkMetaFragmentContainer = createFragmentContainer(MyCollectionArtworkMeta, {
   artwork: graphql`
     fragment MyCollectionArtworkMeta_artwork on Artwork {
       title
+      artistNames
       date
       medium
-      dimensions {
-        in
-      }
+      category
+      height
+      width
+      depth
+      metric
     }
   `,
 })
