+++ src/lib/Scenes/MyCollection/Screens/ArtworkDetail/MyCollectionArtworkDetail.tsx
@@ -3,40 +3,26 @@ import {
   MyCollectionArtworkDetailQueryResponse,
 } from "__generated__/MyCollectionArtworkDetailQuery.graphql"
 import { FancyModalHeader } from "lib/Components/FancyModal/FancyModalHeader"
-import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import { ScreenMargin } from "lib/Scenes/MyCollection/Components/ScreenMargin"
 import { AppStore } from "lib/store/AppStore"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
-import { useScreenDimensions } from "lib/utils/useScreenDimensions"
-import { Button, Join, Separator, Spacer, Text } from "palette"
+import { Button, Join, Separator, Spacer } from "palette"
 import React from "react"
 import { ScrollView } from "react-native"
 import { graphql, QueryRenderer } from "react-relay"
-import { MyCollectionArtworkInsightsFragmentContainer as ArtworkInsights } from "./Components/MyCollectionArtworkInsights"
+import { MyCollectionArtworkInsightsFragmentContainer as ArtworkInsights } from "./Components/ArtworkInsights/MyCollectionArtworkInsights"
+import { MyCollectionArtworkHeaderFragmentContainer as MyCollectionArtworkHeader } from "./Components/MyCollectionArtworkHeader"
 import { MyCollectionArtworkMetaFragmentContainer as ArtworkMeta } from "./Components/MyCollectionArtworkMeta"
-import { WhySellStep } from "./Components/WhySell"
+import { WhySell } from "./Components/WhySell"
 
 export interface MyCollectionArtworkDetailProps {
   artwork: NonNullable<MyCollectionArtworkDetailQueryResponse["artwork"]>
 }
 
 const MyCollectionArtworkDetail: React.FC<MyCollectionArtworkDetailProps> = ({ artwork }) => {
-  const dimensions = useScreenDimensions()
   const navActions = AppStore.actions.myCollection.navigation
   const artworkActions = AppStore.actions.myCollection.artwork
-
-  // FIXME: UI fill in props
-  const artworkProps = {
-    demand: "Strong demand",
-    estimate: "$4,500 - $445,000",
-    image: {
-      url: "",
-    },
-    ...artwork,
-  }
-
-  const formattedTitleAndYear = [artworkProps.title, artworkProps.date].filter(Boolean).join(", ")
 
   return (
     <ScrollView>
@@ -44,53 +30,17 @@ const MyCollectionArtworkDetail: React.FC<MyCollectionArtworkDetailProps> = ({ a
         leftButtonText=""
         rightButtonText="Edit"
         onRightButtonPress={() => {
-          artworkActions.startEditingArtwork(artworkProps as any) // FIXME: remove `any` type
+          artworkActions.startEditingArtwork(artwork as any) // FIXME: remove `any` type
         }}
         hideBottomDivider
       />
       <Join separator={<Spacer my={1} />}>
-        <ScreenMargin>
-          <Text variant="largeTitle">{artwork?.artistNames}</Text>
-          <Text variant="subtitle" color="black60">
-            {formattedTitleAndYear}
-          </Text>
-        </ScreenMargin>
-
-        <OpaqueImageView
-          // TODO: figure out if "normalized" is the correct version
-          imageURL={artworkProps?.image?.url?.replace(":version", "normalized")}
-          height={300}
-          width={dimensions.width}
-        />
-
+        <MyCollectionArtworkHeader artwork={artwork} />
         <ArtworkMeta artwork={artwork} />
-
         <Separator />
-
         <ArtworkInsights artwork={artwork} />
-
         <Separator />
-
-        <ScreenMargin>
-          <Join separator={<Spacer my={1} />}>
-            <Text variant="title">Interested in selling this work?</Text>
-            <WhySellStep
-              step={1}
-              title="Simple Steps"
-              description="Submit your work once, pick the best offer, and ship the work when it sells."
-            />
-            <WhySellStep
-              step={2}
-              title="Industry Expertise"
-              description="Receive virtual valuation and expert guidance on the best sales strategies."
-            />
-            <WhySellStep
-              step={3}
-              title="Global Reach"
-              description="Your work will reach the world's collectors, galleries, and auction houses."
-            />
-          </Join>
-        </ScreenMargin>
+        <WhySell />
 
         <ScreenMargin>
           <Button size="large" block onPress={() => navActions.navigateToConsign()}>
@@ -137,6 +87,7 @@ export const MyCollectionArtworkDetailQueryRenderer: React.FC<{ artworkID: strin
               cm
             }
 
+            ...MyCollectionArtworkHeader_artwork
             ...MyCollectionArtworkMeta_artwork
             ...MyCollectionArtworkInsights_artwork
           }
